Completed leaving a review
Added a model that represents the json that we need to send to the end point to create a review. Made sure if the user is not logged in, to be logged in, and if logged in I show if the user can leave a reivew, or has already left a review.