Added loadReviews function to BookCheckoutPage
Added a new function that fetchs based on a pre-set size , now is 5, and used the function on the See more review. On each click 5 books, or if less than 5 remaining reviews will be fetched and added to the div with id of reviewDiv. In the function I iterate over each fetched reviews create a dive element, use createRoot and render to render one by one the review components to the div as new divs. In case the totalNumberOfReviews is less than the pre-set per load number, I don't show See more Reviews button, in case totalNumberOfReviews is equal to totalNumberOfLoadedReviews I show to the top of the page, if the user clicks they will be navigated to the top of the page