Added history related configuraiton
Added a model of history to consume the endpoint from backend server. Created a HistoryShelf component that fetchs the checkout history books. Added a bookCard similar design to it. Also added pagination to it. I added getUsername function in Authenticated file that returns the username in the jwt token used it in the request url since its needed to fetch the books by. Finally added the historyShelf to the ShelfPage