Added Admin page and admin message component
The admin page is accessible only to those users with authorities ROLE_ADMIN. I added the link to footer and navbar. The admin page contains three compoents, today I completed messages where the admin can see only pending messages and can response. Added in Authenticated a method through which I get the authorities of the user to make the admin page secure. I added a model of AdminresponseMessage with corresponds to the json the backend server expects when using the admin response endpoint. Lastly I used npm bootstrap and installed types for bootstrap in case I used it in future.