HTTPS key and certificate and rleated configuration
In package.json enabled the app to be HTTPS. Used openssl to create a self-signed certificate and key. Create .env file that points to the key and certificate. Finally since I configured the backend server to be secure as well, I add REACT_APP_API property in the .env file which is base url for the backend and injected that in every component that makes a request to the server.