Added Payment page and stripe configuration
First I added in index.html script link to js.stripe.com which we need to connec to stripe. Second In index.tsx I used loadStripe from stripe which accepts the publishible key, and Used in in Element from stripe to wrap my app in to be available through out the app. I created a PaymentInfo model that stripe api accepts when creatig a paymentIntent. I created A PaymentPage that shows the fees that the user has to apy and included CardElement from stripe to handle the sensible card details. Finally in an async function Based on the fees I request the backend server to create the paymentIntent and if everything was successfull I proceed to completing the payment by making another reuqest to the server to remove the outstanding fee. In case of error or card details error I show a warning and in case of success I show the user that payment was successfull and zero their fees. In BookCheckoutPage I have made sure that the user cannot check any book if they have outstanding fees or if they have over-due books to be returned.